requirements.txt
python-dotenv
polars
plotly
notion-client
woob
prettytable
ipykernel
fastapi
uvicorn
flask
python-multipart
ngrok